Fixed price_per_command and price_per_applied_filter with fractional costs
The if case before adding the command is only checking if total > 0. If the total turned out to be 0, then it would be setting the unit's gold cost to 0.

There is an edge case where the cost per command point or filter is fractional, such as #price_per_command #hp 0.25 10, where if the hp over the threshold is low enough, it will get rounded to 0, and thus have the unit cost set to 0.