Refactored debug mode and added some debug logging to ChanceIndHandler
The ChanceIncHandler debug logs help identify when a specific chanceinc or themeinc reduce a filter's chance of occurring down to 0 from a previously higher amount.

This is useful to help debug when certain units end up without a single possible piece of equipment or valid filter, as the debug log will typically be the last thing that displays before an exception is thrown.