Mount item now overrides mount form sprite. Added #barding support too
If a mount item (i.e. an item with #mountmnr) has a #sprite property, this will override the mount form's #spr1 and #spr2 commands.

In most cases, we expect the sprite of the mount that the rider is riding on to be the same as when the rider is dismounted and only the mount is left. This also allows us to define the sprite for the barded mounts in one place, rather than having to duplicate all paths in the mounts.txt list of mount shapes and within the list of mount items too.