More mounts and shapes template updates
Split the mounts.txt and secondshapes.txt files into a separate index file to follow the ResourceStorage pattern with its load() file. These index files then load named lists of items; i.e. default_mounts, default_secondshapes, etc. The goal here is to create a homogenous pattern for loading data files, as seen in the ResourceStorage.load() calls in NationGenAssets.java.

Edited mount item templates to declare their own #barding directly (this is a special command that's equivalent to #armor, except for the fact that the armor will get assigned to the mount, not the rider).