Intense rework of mounts, mountUnits, and unit id resolution
This aims to facilitate declaring mount templates by extracting the barding and sprite aspects of the mount and making it a responsibility of the actual mount item that gets equipped on the unit's mount slot.

This enables us to only define a single base mount with its Dominions stats, without having to repeat those stat blocks 12 times for 12 different bardings, and then leave the variations of theme, barding and spriting to the actual mount items that each race/pose has in their pool.

A Unit also does not contain a Mount reference anymore. This relationship has been clarified better.

Unit <-> MountUnit -> Mount

A Unit (the cavalry unit / rider) contains a MountUnit (the middle-man which describes the mount as a Unit), which in turn contains a Mount (the pure mount creature whose stats are defined in the mounts files).

A Unit may not contain a MountUnit, but a MountUnit will always contain a reference to its parent Unit (i.e. its rider).

The mount line writing is now also passed onto the Unit for it to call it as part of its own line writing calls.