Added extra checkbox options in UnitCommandRestriction
"Match all commands" forces the restriction to only accept units which have all of the listed commands (acting as a logical AND), rather than the default OR behaviour of accepting a unit as long as it's got one of them.

"Check mount commands" will make the restriction dive into the unit's mount commands themselves. This is necessary to be able to restrict cavalry unit generation to a unit with a specific mount and barding, now that these are no longer combined entities (the regular mounts themselves don't contain the barding; the item equipped to the cavalry pose does).