Fixed hasDominionsId check to safely deal with unresolved custom ids
Custom item ids such as "meteoric_barding" will be an alphabetical string that cannot be parsed with parseInt() in that it will cause it to throw an exception.

The main purpose of this check is to ensure we are dealing with an item that's not tagged as a "sprite only" item that's not an actual piece of equipment ingame, such as a shadow, hands, basesprite, etc.