Removed popping of slot in Unit.setSlot()
The popping of the specific slot being equipped was added to ensure that the slot was "clean" before a new item was equipped into it.

However, each slot in the SlotMap class is essentially a stack, and as it turns out, this is by design. Items in a slot are not meant to be immediately removed because they might have been part of an ItemDependency.

If the previous item is removed willy-nilly, and it turns out it was part of a dependency, the chain of dependencies can break down.

The slots are a stack because only the top element of the stack is "relevant" to the program flow. If that item gets removed, we can go back to the previous state the slot was in before without breaking the chain of dependencies.