Grotesque: Fixed assignment of two-handed weapons like maul to offhand
Because of the underlying armUnit() workflow that automatically tries to assign a weapon slot to units, the cyclops maul being part of the offhand slot could sometimes get undermined by the armUnit assigning a dfferent weapon that does not have a free hand slot.

Now left/right arm weapons have been restructured into weapon and offhand weapons, and the left/right arm distinction is made through #types.

A side effect is that there's no longer a need to call #generateitem 1 offhand on the grotesques pose, since all weapon slot items are now using #needs dependencies to grab the correct offhand, whether that be a left or a right arm.