Fixed itemslot calculation from itemTags - two or more slots would calculate wrong
Because the handleModifier() function already handles an addition to the current value (i.e. hand +1), using the addition assignment of its result, such as hand += handleModifier() would end up adding it too many times.

For example, if two items had hand +1, the for loop would assign 1 to the hand on the first iteration, and then 3, because it would calculate the result of another hand +1 as 2 in the handleModifier(), and then add-assign it again to 1.